// typescript/models/RetrieveNewsArticlesByIds200Response.ts
/**
 * World News API
 * The world\'s news wrapped into a single API.
 *
 * OpenAPI spec version: 2.2.0
 * Contact: mail@worldnewsapi.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RetrieveNewsArticlesByIds200ResponseNewsInner } from '../models/RetrieveNewsArticlesByIds200ResponseNewsInner';
import { HttpFile } from '../http/http';

export class RetrieveNewsArticlesByIds200Response {
    'news'?: Array<RetrieveNewsArticlesByIds200ResponseNewsInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "news",
            "baseName": "news",
            "type": "Array<RetrieveNewsArticlesByIds200ResponseNewsInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RetrieveNewsArticlesByIds200Response.attributeTypeMap;
    }

    public constructor() {
    }
}