// typescript/models/RetrieveNewspaperFrontPage200Response.ts
/**
 * World News API
 * The world\'s news wrapped into a single API.
 *
 * OpenAPI spec version: 2.2.0
 * Contact: mail@worldnewsapi.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RetrieveNewspaperFrontPage200ResponseFrontPage } from '../models/RetrieveNewspaperFrontPage200ResponseFrontPage';
import { HttpFile } from '../http/http';

export class RetrieveNewspaperFrontPage200Response {
    'frontPage'?: RetrieveNewspaperFrontPage200ResponseFrontPage;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "frontPage",
            "baseName": "front_page",
            "type": "RetrieveNewspaperFrontPage200ResponseFrontPage",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RetrieveNewspaperFrontPage200Response.attributeTypeMap;
    }

    public constructor() {
    }
}