// typescript/models/SearchNewsSources200Response.ts
/**
 * World News API
 * The world\'s news wrapped into a single API.
 *
 * OpenAPI spec version: 2.2.0
 * Contact: mail@worldnewsapi.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { SearchNewsSources200ResponseSourcesInner } from '../models/SearchNewsSources200ResponseSourcesInner';
import { HttpFile } from '../http/http';

export class SearchNewsSources200Response {
    'available'?: number;
    'sources'?: Array<SearchNewsSources200ResponseSourcesInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "available",
            "baseName": "available",
            "type": "number",
            "format": ""
        },
        {
            "name": "sources",
            "baseName": "sources",
            "type": "Array<SearchNewsSources200ResponseSourcesInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SearchNewsSources200Response.attributeTypeMap;
    }

    public constructor() {
    }
}